bug fix + inlining more functions
"Okay try this version.  I've defined the "speedy" functions
inline using defsubst and also added a nascent column definition
verifier starting with testing the threshold-guess-tolerance
either nil or integer > -1.

This version caps the estimate to the length of the input
string.  It confuses me that the estimate would ever be bigger
than the string but at least it won't signal."